fix: regenerate as 0016 migration and split sequence-elements server module
- Regenerate the sequence_elements table migration as 0016 (renumbered after merging main's 0015_familiar_mongoose).
- Move promoteTempElements + deriveTokenFromFilename out of src/functions/sequence-elements.ts to src/lib/sequence-elements/. Their non-handler exports were keeping triggerWorkflow alive in the client bundle and tripping the TanStack import-protection-plugin (server-only @tanstack/react-start/server pulled into the elements route chain).
- After the split, src/functions/sequence-elements.ts only exports createServerFn handles whose bodies get stripped from the client bundle.

#271
