+++ src/functions/sequence-elements.ts
@@ -1,8 +1,9 @@
 import { getEnv } from '#env';
-import { moveFile, getSignedUploadUrl } from '#storage';
+import { getSignedUploadUrl } from '#storage';
 import { generateId } from '@/lib/db/id';
 import { ulidSchema } from '@/lib/schemas/id.schemas';
-import { STORAGE_BUCKETS, getPublicUrl } from '@/lib/storage/buckets';
+import { deriveTokenFromFilename } from '@/lib/sequence-elements/derive-token';
+import { STORAGE_BUCKETS } from '@/lib/storage/buckets';
 import {
   getExtensionFromUrl,
   getMimeTypeFromExtension,
@@ -14,19 +15,6 @@ import { createServerFn } from '@tanstack/react-start';
 import { zodValidator } from '@tanstack/zod-adapter';
 import { z } from 'zod';
 import { authWithTeamMiddleware, sequenceAccessMiddleware } from './middleware';
-
-/**
- * Derive an uppercase token from a filename stem.
- * Strips extension, non-alphanumeric characters, collapses runs to `_`.
- */
-export function deriveTokenFromFilename(filename: string): string {
-  const stem = filename.replace(/\.[^.]+$/, '');
-  const cleaned = stem
-    .toUpperCase()
-    .replace(/[^A-Z0-9]+/g, '_')
-    .replace(/^_+|_+$/g, '');
-  return cleaned.length > 0 ? cleaned : 'ELEMENT';
-}
 
 /**
  * Ensure a token is unique within a sequence. Appends `_2`, `_3` if taken.
@@ -159,89 +147,6 @@ export const finalizeElementUploadFn = createServerFn({ method: 'POST' })
 
     return element;
   });
-
-// ============================================================================
-// Promote temp uploads (created before sequence existed) into a new sequence
-// ============================================================================
-
-const tempUploadSchema = z.object({
-  tempPath: z.string().min(1),
-  tempPublicUrl: z.string().url(),
-  filename: z.string().min(1),
-});
-
-export async function promoteTempElements(params: {
-  scopedDb: import('@/lib/db/scoped').ScopedDb;
-  teamId: string;
-  userId: string;
-  sequenceId: string;
-  uploads: Array<z.infer<typeof tempUploadSchema>>;
-  triggerVision?: boolean;
-}): Promise<void> {
-  const {
-    scopedDb,
-    teamId,
-    userId,
-    sequenceId,
-    uploads,
-    triggerVision = true,
-  } = params;
-  if (uploads.length === 0) return;
-
-  for (const upload of uploads) {
-    const tempPrefix = `elements/${teamId}/temp/`;
-    if (!upload.tempPath.startsWith(tempPrefix)) {
-      console.warn(
-        '[promoteTempElements] Skipping non-temp path:',
-        upload.tempPath
-      );
-      continue;
-    }
-
-    const relativeTempPath = upload.tempPath.slice('elements/'.length); // teamId/temp/xxx.ext
-    const ext = getExtensionFromUrl(upload.tempPath);
-    const newId = generateId();
-    const permanentRelative = `${teamId}/${sequenceId}/${newId}.${ext}`;
-    const permanentPath = `elements/${permanentRelative}`;
-
-    if (getEnv().E2E_TEST !== 'true') {
-      await moveFile(
-        STORAGE_BUCKETS.ELEMENTS,
-        relativeTempPath,
-        permanentRelative
-      );
-    }
-
-    const publicUrl =
-      getEnv().E2E_TEST === 'true'
-        ? upload.tempPublicUrl
-        : getPublicUrl(STORAGE_BUCKETS.ELEMENTS, permanentRelative);
-
-    const rawToken = deriveTokenFromFilename(upload.filename);
-    const token = await ensureUniqueToken(scopedDb, sequenceId, rawToken);
-
-    const element = await scopedDb.sequenceElements.create({
-      id: newId,
-      sequenceId,
-      uploadedFilename: upload.filename,
-      token,
-      imageUrl: publicUrl,
-      imagePath: permanentPath,
-      visionStatus: 'pending',
-    });
-
-    if (triggerVision && getEnv().E2E_TEST !== 'true') {
-      await triggerElementVision(
-        element.id,
-        sequenceId,
-        element.imageUrl,
-        element.uploadedFilename,
-        teamId,
-        userId
-      );
-    }
-  }
-}
 
 // ============================================================================
 // List / delete / rename
