+++ src/functions/sequences.ts
@@ -34,7 +34,7 @@ import { createServerFn } from '@tanstack/react-start';
 import { zodValidator } from '@tanstack/zod-adapter';
 import { z } from 'zod';
 import { authWithTeamMiddleware, sequenceAccessMiddleware } from './middleware';
-import { promoteTempElements } from './sequence-elements';
+import { promoteTempElements } from '@/lib/sequence-elements/promote-temp-elements';
 
 export const getSequencesFn = createServerFn({ method: 'GET' })
   .middleware([authWithTeamMiddleware])
